Add tests for full forgotten passphrase process
We were only testing that a user could request a password reset and not that they could then
reset their passphrase.

The one piece of the (successful) process that this doesn't test is whether the link in the
email is correct. I decided to forego parsing the HTML from the email for now.