An app can request a user's details even if they don't have permissions in that app
Previously if a user tried to sign in to an app they don't have any permissions in
an exception would be raised when that app tried to store the last updated time of
their permissions. We will now return a response without any permission details and
trust the app's to behave accordingly.