Disable editing of application name
We need to do this because the permissions emitted by Signon include
the application's name. Each application finds the relevant permissions
by looking for it's name as configured. If that is changed in Signon,
it would need to be changed in the app's config (and any cached data),
otherwise everybody's permissions would effectively disappear.

We can revert this when we have made all the apps to support both the
current data structure AND support a new one that excludes the
app's name.