Fix errors when rendering Doorkeeper views
This started happening because uses of this application's path
and view helpers were added to application.html.erb, but because
Doorkeeper is a Rails engine, it doesn't have (direct) access to
these things by default. It's worth noting that the same isn't true
of Devise, but that's because Devise cheats with a monkeypatch.

http://stackoverflow.com/questions/7588870/engine-routes-in-application-controller#answer-7622755