Fix crash for some users accepting invites
For reasons we couldn't determine, when a specific user tried to
accept, they would get an error inside of "confirm!" as Devise tried
to check whether the confirmation token (note that it is not the
invitation token) had expired.

We could only reproduce the problem by manually creating the record
(via SQL, not the model) and then attempting the confirmation flow.
Probably it is some subtle difference in their record, that has come
about because it was created before the migration.

The reproduced problem was fixed by this code change.

Given that in an invitation acceptance flow, you aren't confirming,
it doesn't really make sense to "confirm". Let's manually do what we
invoked that method to do for us: set confirmed_at, thus allowing
the user to login (confirmable, when activated, adds that requirement).