Remove app name from user permissions
This is the last piece in the puzzle for removing the application name from
permissions when they are passed to apps in the OAuth requests (and permission
push).

It means that instead of emitting a hash of app names to arrays of permissions
(which for a long time have only ever included the app making a request or
receiving the push notification), we will send a simple array.

This means that the app doesn't need to know the name used in signon, and we
can allow it to be edited (in Signon) freely.