Make user name consistently not nullable
User names haven't been nullable for some time. Let's ensure that the
databases are consistent.

The production, staging and preview databases all have user name set
to not nullable, but if you migrate a database from scratch it is the
reverse.

This isn't generally a problem as we clone from production, but it would
be nice if it was consistent. It has caused some occasional mismatches
when a dev has built a database from migrations, and then the schema.rb
has changed the nullability of the user name.

This migration shouldn't change anything in production, but will ensure
that it is not nullable everywhere, rather than because we happen to build
our databases in a certain way.