Add the signin permission to the SSO Push User
Currently the SSO Push User only has the `user_update_permission`
permission, however due to GDS::SSO not skipping the requirement
for the `signin` permission explicitly for the push requests,
there is inconsistent behaviour with each app depending on whether
the `signin` permission is required at the level of the
`ApplicationController`.

The easiest way to fix this right now is to give the SSO User
the `signin` permission too. We may want to make a change to the
SSO gem in the future to not require this.