Add UI for editing user membership of organisations
This needs to be hidden behind a feature flag until Signon becomes
the canonical source for user membership of organisations. Until
then, these relationships should be edited in Whitehall.

Abbreviations need to be displayed as well as names so that users can
find organisations by typing the abbreviation, but the abbreviation
needs to come after the name because not all organisations have
abbreviations, and some are long. The abbreviation also needs to have
a space before it for finding by typing to work, but putting it in
parentheses with spaces looks odd, so it is separated from the name
with a | when it's used.