Remove closed_at datetime from organisations
Not all closed organisations in Whitehall have a closed_at datetime
set (2 out of 4 do have it) so it's not reliable for finding closed
organisations. If we do want to provide different behaviour in future
for closed organisations or those which are exempt from transitioning
to GOV.UK, govuk_status from Whitehall is the most accurate available
source for that information.