Don't show user_update_permission in user form
Remove it from the supported_permission_strings list so that it doesn't appear
in the GUI. This was introduced in: 505a51a93cbb4d60ccc857a8e21b5b00ccb43242

We moved the check for unsupported permissions to the rake task so that
SSOPushCredential could still use grant_permissions to add the (supposedly)
unsupported permission.

This was easier than trying to make the supported permissions list vary based
on whether or not the user is the SSOPushUser.