Change user-organisation relationship to many-to-one
We initially chose to create a many-to-many relationship between
users and organisations because we thought that we would need that
flexibility when dealing with access controls for the Transition app.

After further discussions about Whitehall's existing use of user
membership of (single) organisations, and after thinking more about
what we need for Transition, we have decided that we have no
immediate use case for this being a many-to-many relationship and that
to do it that way at this stage introduces unnecessary complexity for
Whitehall and Transition.