Rename organisation to organisation_slug
When it came to making use of this data in an application, it quickly became
clear that it was going to be common for people doing the same task to want a
method on their user called 'organisation' which would clash directly with this
field name.

We are still in a position where we can change the field name to reflect what
it really is and to make integration for apps cleaner and simpler.

If we were ever to change how we passed around organisation references (eg to a
full URL) then we would need to put it in a different field name anyway.