Fix test that was spuriously passing with sqlite3
sqlite3 does case-sensitive sorting, which means that "email" was
sorting before "A name is now...", and therefore when no letter was
supplied, the index was defaulting to 'E'.  With the switch to mysql,
this broke the test because it was now defaulting to 'A'.

See
https://github.com/edendevelopment/paginate_alphabetically/blob/master/lib/paginate_alphabetically/active_record.rb#L16-L20
for the code that does this.