Swapping alphabetical pagination gem
We need the alphabetical pagination logic to take
care of the queries containing a join for the upcoming
organisation admin role. there is a need to scope users by
the first letter as well as the organisation they belong to.

The earlier gem doesn't allow us to specify something like
users.name, which we will need when we join with organisations,
or else mysql gets confused.