Bugfix for triggering remote requth on logout.
Warden::Manager.before_logout was being triggered when visiting signon
when your session had expired.  This had the effect of causing you to be
remotely signed out of all apps that you were using.

Moving the ReauthEnforcer call into the sessions controller means that
it only gets triggered on an explicit logout.