Update user json to reflect user suspension
www.agileplannerapp.com/boards/173808/cards/5886

Add a disabled flag to the user json that gets
synced with the user's apps, so that they can
treat suspended users differently.

Exclude permission information to be sent for a
suspended user, because the suspended user should
not be able to perform any operation. Whenever the
user's account gets un-suspended we can send the
correct permissions back to the user's apps.