Signed-in users with expired password can reset it
www.agileplannerapp.com/boards/173808/cards/8676
fixes: https://govuk.zendesk.com/agent/tickets/874886

by default `require_no_authentication` filter will redirect a
partially signed-in user to the after sign-in path, i.e. '/'
where expirable will intercept and ask the user to change
their password if it has expired.

in case of the zen ticket mentioned above, a user with an
expired password is unable to reset it, because clicking on
the reset password link takes them to the change password
screen asking them to input their current password (which
they don't remember).

we're signing-out the user in this particular situation
when they click on the reset password link, so that they're
straight-away taken to the reset password screen.