Indicate which environment an email is from
Include the environment name in the From field, unless it's production.
instance_name isn't set in production.

Currently the only way to distinguish which environment an email is from is by
links in the email.

This is a problem for many users as they have accounts on production and
preview which they use on a daily basis.

In the same way that the app displays which environment it is running in,
emails should include the environment they came from.