Migration to add permissions join table
The existing permissions table stores permissions as a string
field which contains permissions as a serialised array. This
has made it difficult for us to write reporting scripts in
the past and complicates code.

It also stores one row per application per user, which is not
necessary as a user may not have access to all applications.

We're moving this data to a join table between user and supported
permissions which is the more appropriate way to model such data.
This will also allow rails to clean up records for supported
permissions that no longer exist.