Destroy user permissions with supported permission
in the earlier way in which we stored permissions (as serialised strings)
it was impossible to dependent destroy without writing hooks that would
have to scan through all permission records and modify the serialised
array.

with the new permissions model, we can use rails' `dependent: :destroy` to
achieve the same.