Move batch invitation permissions to a join table
a batch invitation specifies the permissions to be assigned to
users being invited, which was stored as a hash in a serialised
attribute.

we've replaced that serialised attribute with a join table, so
that it is modeled the same way as user application permissions.

there is no harm in removing application_and_permissions attribute
from the model after existing batch invitations have been processed.
only while creating and processing a batch invitation we need that field.