Improve passphrase reset tests to use email links.
These now test the full flow including following the links in the email.
This provides more complete coverage of the user journey.  This also
adds a test to ensure that a password reset link can't be used more than
once.

This required changing the action_mailer host for the test environment
to match the one Capybara uses.  Without doing this links followed from
emails use a different session because the cookies are tied to the
hostname.