Don't change reset_password_token on first access.
This partially reverts a044b79 (#65).

It would seem that some of our users have virus-scanning systems that
follow links in emails to check for anything malicious.  This was
breaking the passphrase reset flow because the token was being reset the
first time the page was accessed.  The root of this problem is that we
have a GET route in the application that has side-effects.

This removes this functionality so that a password reset token will
continue to work until the passphrase has been reset, at which point
it's cleared (this is standard devise behaviour).