Make the user's name a link to edit themselves
It's quite a common use case, at least for GOV.UK developers, to want to
change their own permissions to test some permission-based feature in an
app, which involves going to the user section and searching for their
user record. If they have to do this multiple times, it's tedious and
introduces the risk that they'll change permissions for someone else by
mistake.

The original destination for this link was to the Signon root page: this
is consistent with other apps, where it makes sense to link through to
Signon; when you're in Signon already, this isn't particularly useful,
especially when the "GOV.UK Signon" link in the header takes you there
already.

The logic to determine the link target depends on whether the user has
permission to edit themselves. As the logic currently stands, every user
in the system has permission to edit themselves to at least some extent,
but we shouldn't assume this will always be the case.