Force usage of SSL in redirect URIs
All our existing redirect URIs are https.

https://github.com/doorkeeper-gem/doorkeeper/pull/489

"Since OAuth2 does not implement any encryption itself and relies on
SSL for transport security, it is wise to use HTTPS in communication
instead of HTTP."

Doorkeeper enforces this if force_ssl_in_redirect_uri is true, which
is false by default in RAILS_ENV development.