Drop permissions table, migrated to join table
following this change:
https://github.com/alphagov/signonotron2/pull/311
user's application permissions are migrated join table instead of
being stored as a serialised attribute in the permissions table.

dropping the permissions table which remained for backup. now that
we're running with the new code for a week and nothing's broken
we should be good to clean that up.