Fix auto-generated user update permissions
after saving an application that supports push updates,
we create a supported permission 'user_update_permission'
for it. this was being created such that it was visible
on the Signon interface, whereas it shouldn't have been.

the issue has been fixed, and this migration cleans-up
any invalid data that's been created in the meantime.