Bugfix - viewing user with unsynced permissions
In some circumstances a crash could occur when viewing a user with unsynced
permissions. The precise scenario in which this could occur is if the first
permission in the users permission list was unsynced, and later ones were
synced.

In theory this should never occur because all permissions are synced at
the same time however the association from user to application_permissions is
unordered, so the order in which the records are returned from mysql when
reading them could change arbitrarily.

The fix here is to make the display code more robust to this possible scenario.

I've also refactored the view slightly so that it's clear that only the
`<span>` label class changes according to the `sync_needed?` status.

https://www.pivotaltracker.com/story/show/88500338