Fix the Devise session extension bug in place
There is a security hole in versions of Devise prior to 3.2.0, in which
a logged-in user sending a GET request to `/users/sign_in` would have
their session extended, even if it had already expired; this also
doesn't record anything in the event log, as the user hasn't technically
signed in again. As a side-effect of this, people who have ended up
using that URL as their bookmark for Signon get sent suspension
warnings, as we don't have a record of them having logged into Signon.

Given the number of Devise plugins we use, and the customisations we've
made on top of them, upgrading to a patched version of Devise is
sufficiently complicated that it's worth patching this version in place
in the meantime, so we aren't vulnerable while working on the upgrade.

The commit in Devise that fixes this issue is
<https://github.com/plataformatec/devise/commit/52a3768451e4ef1e>: it
restricts a filter so that it doesn't ignore session timeouts when
displaying the login form, while still ignoring them when signing in or
out. Ideally, we'd be able to throw in a `skip_before_filter` and be
done, but this filter is defined as an anonymous block, so we need to do
some messing around in the guts of the callback chain to find the name
Rails has assigned to the callback we're after.

This piece of code depends heavily on the order in which Devise defines
the callbacks, so is locked to the version we're currently using, and is
moderately paranoid when checking that the callback looks like the one
we want: if anything isn't as expected, the patch will abort messily and
break the app. This is a deliberate decision, so we can't miss it and
cause unknown problems when we upgrade Devise itself. It's also not been
factored for readability, because the very next change to this app
should be to upgrade Devise so that this code is no longer needed.