Changed .ruby-version file in order to make use of the latest Ruby version (2.2.2) Starting from this version some old native gems stop compiling. Because of this we had to update unicorn and its related gems (kgio and raindrops) as well as json gem.
In addition, the app was using an outdated custom fork of devise_security_extension. The gems dependencies caused an infinite loop bug when we tried to run any bundle update command.
Since the issues that made us create the custom fork were solved, we switched back to the release version of the gem where the infinite loop bug was solved.