Clean up orphaned organisations
We have 51 organisations in production and 84 in preview which do not exist in
Whitehall's Organisation API. This is because they either changed slug, and we
created a duplicate with the new slug, or they have been deleted from
Whitehall.

Other than relating them to users, we don't do anything with organisations, so
it seems safe to delete them.

I have gone through and reassigned the small number of users assigned to these
orphaned organisations in both environments, so the migration should complete
first time.

In development, I've set it to simply clear out the value as it doesn't really
matter and developers will get an up to date data from preview the next time
they sync their data.