Pin Rake to 10.4.1
This is to stop the --trace option to Rake being passed to Test::Unit,
which then raises an invalid option error for it, breaking the build.
ARGV handling changed in 10.4.2, which appears to be responsible for this.

I haven't been able to reproduce this error in dev, which makes it hard to
track down exactly what's going on; `bundle exec rake --trace` runs fine
for me there on 10.4.2, but it has failed in CI multiple times on 2
separate machines.