Display when organisations have closed
This makes it clearer when choosing organisations to assign to a user which
ones are plausible. An example of this is Highways Agency which has become a
different organisation called Highways England.

This could happen more when the new government is formed and organisations are
closed and created with potentially similar/confusing names.