Make password expiry tests more robust
These tests started failing today without any changes. This seems to be because
the tests were setting the password to have expired `3.months.ago` and Devise
`expire_password_after` is set to `90.days`, which is usually but not always
the same.

Today, this is what Rails returned:
```
Loading development environment (Rails 3.2.18)
irb(main):001:0> 3.months.ago
=> Sun, 01 Feb 2015 14:24:23 GMT +00:00
irb(main):002:0> 90.days.ago
=> Sat, 31 Jan 2015 14:24:26 GMT +00:00
```

We think that the intention was for it to set the password_changed_at to be
further in the past than the expiry time, so we changed it to 91 to make it
clearer.