Re-add build_resource to InvitationsController
devise_invitable still expects the build_resource method to exist on the
DeviseController, but it was removed in Devise 3.0:

    https://github.com/plataformatec/devise/commit/8a93c340802c9674cfcc39eec12793bd63cf5776#diff-d87e90fea7bf5b74d9b8c0e27c1316ceL109

There isn't a safe later version of devise_invitable which we can upgrade to
now, so staying on 1.1.5 and adding this method back into this controller
seems like the best way to get to Devise 3.0.

The original build_resource method in Devise checked for an `unsafe` option,
but we aren't using it and can't see it in devise_invitable, so we've added a
simplified version of this method. We also use resource_params instead of raw
params here, which have been sanitised.