Encode `reset_password_token` that was sent to the user
Use it to look for the user in the database instead of the token that was sent. Fix the functional tests to use the correct token too.

Devise 3.1.0 sends a token in an email *and* stores an encoded version of the reset token against the user. See https://github.com/plataformatec/devise/compare/v3.0.4...v3.1.0#diff-fa817bf486c518644016136465d914b2L46