Update Airbrake params filtering to match Production
We already have this params_filters config in the Airbrake initializer in
production environments, so it should be here too for consistency. Including
password-strength-score in the Airbrake filter is done for consistency with
the Rails parameter filtering.

Also add a comment documenting the fact that our existing parameter
filtering for Rails logging already matches all the sensitive password
params.