Fix population of application permissions fields in user form
User.permission_ids_for was accidentally changed in e96e6744a78fcd8b743f6a3e3a46515996067e3e
so that it wasn't returning the existing permissions for the application.
That meant that the existing permissions were not put into the user
edit form, and so were not re-submitted when updating the user, so
all existing permissions were removed.

This commit fixes the query to find the right permission ids for the
application, so that the form is populated correctly and permissions
are preserved.

We were not testing that pre-existing permissions are preserved when
editing a user, so we have expanded test coverage to check that case
(and also check that another supported permission is not given to the
user).