Prevent normal users from changing their email without verification
This flaw could have been used by an attacker to more easily fully takeover an
account, but they would need to get access to the session in either case.

There's also a usability concern here - the user might make a mistake and set
it to a non-existent email address they have no access to. The confirmation
step prevents that kind of change.