Add password archival support
This is a reimplementation (read: stealing and fixing) of the password_archivable functionality from the now-removed devise_security_extension gem.

It continues to store a history of encrypted passwords in the same table as was used by the gem. When a new password is chosen, it checks if it is valid for either of the last two encrypted passwords for that user. If it is, it returns a validation error.