Add password expiry functionality
This is a reimplementation (read: stealing and fixing) of the password_expirable functionality from the now-removed devise_security_extension gem.

It stores the time the password was last changed or created on the user object. On each request it checks if the stored time is more than 90 days ago, and if so forces the user to change their password before they can log in again. It relies on the password_archivable module, without which the user would be able to "change" their password to their existing password.