Don't try to debug log non-existent methods
There was an instance in production where a job raised an EOFError,
which doesn’t have a `#response` method. This caused the
Rails.logger.warn line to fail.

This adds an inner rescue block to catch the case where we try to log
an exception that doesn’t define `#response`, or anything else in the
chain.

In addition, this moves the `notify_airbrake` calls above the log lines
in order to get the full underlying error in case of the logging
failing.

Discovered from
https://errbit.production.alphagov.co.uk/apps/52ef92340da11517d5000003/p
roblems/55bb0f21657863081b5b0000