Speed up reporting user's app permissions
This was slow because each row required looking up the individual
user's permissions and name of the permission for each available
application, all using ActiveRecord associations. We can fetch and
cache all permissions for all users as this is still more efficient
even if the users are being filtered by status etc.

This change brings the request down from ~15 seconds to ~4 seconds
on the development VM.