Reimplement controller logic
Logic from the TwoFactorAuthentication controller in the gem is now in
the TwoStepVerification gem. This also makes the 2SV controller a
direct subclass of DeviseController

Adds common logic into a helper module, and include that in
ApplicationController instead of adding it directly to
ApplicationController.

Scope is hard-coded to `:user` and similarly `resource` is
`current_user` to improve code readability.