Add 2SV session memory
This was missed in the removal of the two_factor_authentication gem
from this repo.

This also improves upon the security of the original implementation by
verifying user id and expiration, making it impossible to copy the
value of the signed gem and re-use it later, extend the expiration
period, or use your cookie for another user.