Test tampering with 2SV session cookie expiry
This tests preserving the 2SV session cookie value past the cookie’s
expiry date, or otherwise modifying the expiry date. If we weren’t
checking the signed `valid_until` value of the cookie, the user would
be able to preserve the cookie and carry on logging in without 2SV
after 30 days.