Don't load simplecov unnecessarily
SimpleCov was running and producing coverage reports when I ran `rake`
in the development environment, which isn't good. Also, it wasn't
really configured properly for CI either.

The question is, does anyone care about coverage? If yes, merge this.

If no, I'll just blow it all away and forget it ever happened :boom: