Add ability for users to change the device used for 2SV
This merges the existing create flow with the update flow, as it’s not
currently possible to add 2SV devices, only replace them. The result is
the update of the user’s OTP key, either from nil to a valid key thus
setting up 2SV, or from an old key to a new one, thus changing the
device they’ve got set up.

The same form is displayed for both cases, the differences are in the
text of the logged event, and the flash messages on success. The flows
are differentiated by the presence of the `otp_secret_key` on the user.

This commit also refactors the controller action responsible for
updating the secret by extracting the logic to verify the entered code
and update the secret.

Microcopy differentiations between the two flows will come in a later
commit