Prompt future admins and superadmins to set up 2SV
https://trello.com/c/Xh7kLY4z/150-prompting-all-current-and-future-admin
s-and-super-admins-to-set-up-2sv-in-production-only-small

When users are created as admins or superadmins, or promoted to those
roles, they should gain the `require_2v` flag. If a user is demoted,
they shouldn’t lose the flag. If an admin has been flagged as not
requiring 2SV and an unrelated attribute is changed, their 2SV status
should remain the same.