Test for correct drift
There will be points where this would have failed with 35 seconds. The drift only allows for 30 seconds, so if this test had run in the first 5 seconds of the 30 second 2SV window, it would fail. This needs to be 30 seconds max, and in future we can look to change the configured drift seconds.