Simple `event` to `event_id` mapper
For all types of event, issue a single `UPDATE` to add the mapped
`event_id`. Despite the size of the production dataset, this should be
reasonably efficient.

This leaves the `event` attribute as-is for now. We can come back and
wipe that out in a separate migration to drop the column.