Auto-flag adminish users for 2SV in production only
A new integration environment is being introduced to eventually replace
preview, so logic is being switched to only auto-flag users in
production environment.

This behaviour should eventually be controlled by a specific
configuration variable in line with 12 factor, but we need to get this
updated quickly in order to work with the new environment.

Additionally, this now depends on `instance_name` directly in line with
other logic in this app. For production, `instance_name` is `nil`.