Show permissions of user
It is helpful to see the permissions of the users for each app.

We can't use the existing `permission_for(application)` here, because
that will kick off two SQL queries for each user in the list. Just
adding eager loading the permissions with `includes` doesn't work,
because it will still do a query. To avoid the N+1 query, we add a new
method that uses the data that's already loaded.