Fix app_title display translation missing
Since c79a514, we were using Rails' I18n to
set the GovukAdminTemplate.app_title in an initializer. Unfortunately, whilst
the I18n code is available, it hasn't been configured at the time intializers
are loaded. This meant that users saw this in the page header:

  translation missing: en.department.name Signon

The tests didn't fail because the setting to make missing translations raise
errors (in config/environments/test.rb) is also loaded after initializers.