Record when a passphrase reset page is visited with expired token
It would be useful to see when users are trying to use expired tokens.

Note that we can't log when a user tries to use a token from a passphrase reset
request that has been followed by another passphrase reset request. This is
because we can't relate the token to a user, because it is overwritten.