Make pagination of users DB agnostic
The `alpha_pagination` gem we're using doesn't support native SQL query
generation when running under the postgresql adapter. It also has a
strange API, requiring a block to be yielded to determine the sort
field. This change ensures that pagination works when running in either
postgresql or mysql configuration.

We should ensure we replace this (seemingly trivial) pagination with
something fully native once we migrate to postgresql proper.