Make tests run agnostic of locale
We have removed the GOV.UK specific text from the application so that we can use pension wise specific text on our server. This was done using a pension wise specific locale file.

In order to test that everything works as expected and all locale text has been populated we want to be able to run the tests suite with our alternative locale file.

As a result, I removed and GOV.UK specific text from the tests and used regular express matches instead.

- Remove reference to "digital.cabinet-office"
- Deal with from address having optional quotes
around the name portion of the email if the name 
contains any special characters.  In this case 
the '.' in 'GOV.UK'